Package Billing Purchase order list Approval

+++ Libraries/HIMS.Data/DTO/IPPatient/IPPackageDetailsListDto.cs
@@ -8,9 +8,9 @@
         public long ServiceId { get; set; }
         public string? ServiceName { get; set; }
         public long? PackageServiceId { get; set; }
-        public double? Price { get; set; }
+        public decimal? Price { get; set; }
         public double? Qty { get; set; }
-        public double? TotalAmt { get; set; }
+        public decimal? TotalAmt { get; set; }
         public double? ConcessionPercentage { get; set; }
         public decimal? ConcessionAmount { get; set; }
         public decimal? NetAmount { get; set; }
